+++ webrtc/modules/interface/module.h
@@ -25,7 +25,7 @@ class Module {
 
   // Returns the number of milliseconds until the module want a worker
   // thread to call Process.
-  virtual int32_t TimeUntilNextProcess() = 0;
+  virtual int64_t TimeUntilNextProcess() = 0;
 
   // Process any pending tasks such as timeouts.
   virtual int32_t Process() = 0;
