+++ webrtc/modules/pacing/include/mock/mock_paced_sender.h
@@ -29,7 +29,7 @@ class MockPacedSender : public PacedSender {
                                 int64_t capture_time_ms,
                                 size_t bytes,
                                 bool retransmission));
-  MOCK_CONST_METHOD0(QueueInMs, int());
+  MOCK_CONST_METHOD0(QueueInMs, int64_t());
   MOCK_CONST_METHOD0(QueueInPackets, int());
 };
