+++ webrtc/test/direct_transport.cc
@@ -78,9 +78,9 @@ bool DirectTransport::NetworkProcess(void* transport) {
 
 bool DirectTransport::SendPackets() {
   fake_network_.Process();
-  int wait_time_ms = fake_network_.TimeUntilNextProcess();
+  int64_t wait_time_ms = fake_network_.TimeUntilNextProcess();
   if (wait_time_ms > 0) {
-    switch (packet_event_->Wait(wait_time_ms)) {
+    switch (packet_event_->Wait(static_cast<unsigned long>(wait_time_ms))) {
       case kEventSignaled:
         packet_event_->Reset();
         break;
